+++ laktory/models/table.py
@@ -54,7 +54,6 @@ class Table(BaseModel):
     # Computed fields                                                         #
     # ----------------------------------------------------------------------- #
 
-    @computed_field
     @property
     def parent_full_name(self) -> str:
         _id = ""
@@ -69,7 +68,6 @@ class Table(BaseModel):
 
         return _id
 
-    @computed_field
     @property
     def full_name(self) -> str:
         _id = self.name
@@ -77,7 +75,6 @@ class Table(BaseModel):
             _id = f"{self.parent_full_name}.{_id}"
         return _id
 
-    @computed_field
     @property
     def schema_name(self) -> str:
         return self.database_name